Rename mywalkdir.go -> myreaddir.go, remove unused functions
